refactoring, splitting rust in modules and make frontend more modular
