+++ src-tauri/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "gsx-profile-manager"
-version = "1.0.2"
+version = "1.0.3"
 description = "A Manager to manage GSX Profiles"
 authors = ["grtn"]
 edition = "2021"
