+++ src-tauri/tauri.conf.json
@@ -1,7 +1,7 @@
 {
   "$schema": "https://schema.tauri.app/config/2",
   "productName": "GSX Profile Manager",
-  "version": "1.0.2",
+  "version": "1.0.3",
   "identifier": "com.gsx-profile-manager.app",
   "build": {
     "beforeDevCommand": "pnpm dev",
